Create SSIS-ScriptComponent-Code.cs
Here we are using SSIS script component as data source in our SSIS data flow task. We read the files content using System.IO File.ReadAllLines and then process the PII by hashing and masking fields that were passed as package parameters to the script. Finally the transformed data is written to the output buffer to be passed to the flatfile destination